refactor: tuono lib.rs structure
